Keep FixedPoolAllocator in param_and_grad_buffer.py
FixedPoolAllocator depends on ParameterGroup (reads .fsdp_unit_id, .dtype,
.params on each group), so it's tightly coupled to that class. Until the
coupling is refactored separately, keep it colocated with ParameterGroup
in param_and_grad_buffer.py rather than relying on a TYPE_CHECKING forward
reference from allocators.py.

Drops FixedPoolAllocator from allocators.py (along with now-unused
defaultdict/nullcontext/List/TYPE_CHECKING imports), re-inserts it in
param_and_grad_buffer.py right after ParameterGroup, and restores the
defaultdict import there.
