+++ megatron/core/distributed/fsdp/src/megatron_fsdp/param_and_grad_buffer.py
@@ -22,7 +22,7 @@ import inspect
 import logging
 import math
 import warnings
-from collections import namedtuple
+from collections import defaultdict, namedtuple
 from contextlib import ExitStack, nullcontext
 from enum import Enum
 from typing import Callable, Dict, List, Optional, Tuple, cast
@@ -34,7 +34,6 @@ from torch.distributed.tensor import DTensor, Replicate, Shard
 from .allocators import (
     NCCL_ALLOCATOR,
     Bucket,
-    FixedPoolAllocator,
     MultiGroupUBRAllocator,
     RotaryBucketAllocator,
     StorageResizeBasedBucketAllocator,
@@ -838,6 +837,192 @@ class ParameterGroup:
     hfsdp_helper_wtbuf: Optional[DataParallelBuffer] = None
     hfsdp_helper_gbuf: Optional[DataParallelBuffer] = None
     hsdp_comm_gbuf: Optional[DataParallelBuffer] = None
+
+
+class FixedPoolAllocator(TemporaryBucketAllocator):
+    """
+    A specialized temporary bucket allocator that implements a buffer recycling strategy
+    to minimize memory fragmentation in FSDP operations.
+
+    This allocator maintains a fixed pool of pre-allocated buffers, reusing them
+    to reduce the overhead and fragmentation caused by frequent allocation and
+    deallocation of temporary buffers during FSDP operations.
+    """
+
+    def __init__(
+        self,
+        name: str,
+        fsdp_param_groups: List[ParameterGroup],
+        size: int = 2,
+        fallback_to_persistent_buffer: bool = False,
+    ):
+        self.name = name
+        self.fsdp_param_groups = fsdp_param_groups
+        self.size = size  # Number of buffers in the pool (default is 2 for double buffering)
+        self.allocation_tracker = {}  # tracking the global buffer allocation status
+
+        # Build a mapping from FSDP unit id to its associated bucket ids.
+        fsdp_unit_buckets = defaultdict(list)
+        for bucket_id, param_group in enumerate(fsdp_param_groups):
+            if param_group.fsdp_unit_id == -1 or param_group.fsdp_unit_id is None:
+                continue
+            fsdp_unit_buckets[param_group.fsdp_unit_id].append(bucket_id)
+        self.fsdp_unit_buckets = fsdp_unit_buckets
+
+        # Identify the largest group of FSDP units that share the same buffer storage.
+        fsdp_units_to_double_buffer = []
+        for fsdp_unit_id, bucket_ids in fsdp_unit_buckets.items():
+            same_storage_fsdp_units = []
+            for i in fsdp_unit_buckets:
+                if self._is_two_bucket_group_equal(fsdp_unit_buckets[i], bucket_ids):
+                    same_storage_fsdp_units.append(i)
+            # Track the largest group of FSDP units sharing the same buffer storage
+            if len(same_storage_fsdp_units) > len(fsdp_units_to_double_buffer):
+                fsdp_units_to_double_buffer = same_storage_fsdp_units
+
+        # --- Fixed Pool Buffering Check ---
+        # Ensure there is at least one group of FSDP units eligible for fixed pool buffering.
+        # If not, the allocator cannot provide its intended memory recycling benefits.
+        assert (
+            len(fsdp_units_to_double_buffer) > 0
+        ), "Found no FSDP units to use fixed-size buffering"
+        self.fsdp_double_buffer_units = fsdp_units_to_double_buffer
+
+        if torch.distributed.get_rank() == 0:
+            for bucket_id, param_group in enumerate(fsdp_param_groups):
+                if (
+                    param_group.fsdp_unit_id == -1
+                    or param_group.fsdp_unit_id is None
+                    or param_group.fsdp_unit_id not in self.fsdp_double_buffer_units
+                ):
+                    logging.info(
+                        f"FSDP unit (id={param_group.fsdp_unit_id}) does not fit "
+                        "in FixedPoolAllcator"
+                    )
+                    if fallback_to_persistent_buffer is False:
+                        logging.info(
+                            "It will fall back to dynamic memory allocator, NCCL user "
+                            "buffer is not supported"
+                        )
+                    else:
+                        logging.info(
+                            "It will be allocated a persistent buffer. If the memory "
+                            "budget is tight, set "
+                            "trainer.strategy.ddp.fsdp_db_use_persist_buf_on_alloc_fail to False."
+                        )
+
+        # Initialize buffer group status.
+        # Each buffer group represents a set of buffers associated with an FSDP unit's bucket group.
+        self.idle_buffer = []  # List of available (buf_group_id, offset) tuples.
+        self.using_buffer = {}  # Map from bucket_id to (buf_group_id, offset) in use.
+
+        # Populate the idle buffer pool with all buffer group and bucket offset combinations.
+        for buf_group_id in range(self.size):  # Iterate over each buffer group in the pool.
+            num_bucket = len(self.fsdp_unit_buckets[self.fsdp_double_buffer_units[0]])
+            for bucket_offset in range(num_bucket):
+                self.idle_buffer.append((buf_group_id, bucket_offset))
+
+        # Fallback allocator used if the fixed pool allocator cannot fulfill a request.
+        self.fallback_to_persistent_buffer = fallback_to_persistent_buffer
+        self.backup_allocator = TemporaryBucketAllocator()
+
+    def _is_two_bucket_group_equal(self, group_a, group_b):
+        # Check if two bucket groups are equivalent in dtype and size.
+        if len(group_a) != len(group_b):
+            return False
+
+        for a, b in zip(group_a, group_b):
+            pg_a = self.fsdp_param_groups[a]
+            pg_b = self.fsdp_param_groups[b]
+            a_size = sum(p.numel() for p in pg_a.params)
+            b_size = sum(p.numel() for p in pg_b.params)
+            if pg_a.dtype != pg_b.dtype or a_size != b_size:
+                return False
+        return True
+
+    def allocate(
+        self,
+        bucket_id: int,
+        size: int,
+        dtype: torch.dtype,
+        device: torch.device,
+        mem_alloc_context: Optional[Callable] = None,
+    ) -> Bucket:
+        """
+        allocate a temporary bucket.
+        """
+        fsdp_unit_id = self.fsdp_param_groups[bucket_id].fsdp_unit_id
+        if fsdp_unit_id in self.fsdp_double_buffer_units:
+            # Try to allocate from the buffer pool.
+            bucket_offset = self.fsdp_unit_buckets[fsdp_unit_id].index(bucket_id)
+            buffer_name = None
+            if bucket_id in self.using_buffer:
+                # If this bucket is already using a buffer, reuse it.
+                buf_group_id, bucket_offset = self.using_buffer[bucket_id]
+                buffer_name = self._get_gbuf_name(buf_group_id, bucket_offset)
+            else:
+                # Otherwise, find an available buffer group for this bucket offset.
+                for buf_group_id in range(self.size):
+                    if (buf_group_id, bucket_offset) in self.idle_buffer:
+                        self.using_buffer[bucket_id] = (buf_group_id, bucket_offset)
+                        buffer_name = self._get_gbuf_name(buf_group_id, bucket_offset)
+                        self.idle_buffer.remove((buf_group_id, bucket_offset))
+                        break
+
+            assert buffer_name is not None, (
+                f"[FSDP][Rank {torch.distributed.get_rank()}][{self.name}] "
+                f"No buffer found for bucket_id: {bucket_id}, fsdp_unit_id: {fsdp_unit_id}, "
+                f"bucket_offset: {bucket_offset} \n"
+                f"current using_buffer: {self.using_buffer} \n"
+                f"current idle_buffer: {self.idle_buffer}"
+            )
+        elif self.fallback_to_persistent_buffer is True:
+            buffer_name = f"{self.name}_not_fit_in_fixed_pool_{bucket_id}_{size}_{dtype}_{device}"
+        else:
+            # If the bucket is not eligible for fixed pool buffering, or no buffer is available,
+            # fall back to dynamic allocation via the backup allocator. This means that we
+            # will do dynamic memory allocation.
+            logging.debug(f"[FSDP] Using backup allocator for {bucket_id} {fsdp_unit_id}")
+            return self.backup_allocator.allocate(
+                bucket_id=bucket_id, size=size, dtype=dtype, device=device
+            )
+
+        # Use buffer_name to get memory from global memory.
+        if mem_alloc_context is not None and mem_alloc_context != nullcontext:
+            # Check if a new buffer allocation is required
+            if (
+                self.allocation_tracker.get((buffer_name, dtype), None) is None
+                or self.allocation_tracker[(buffer_name, dtype)] < size
+            ):
+                # Requires synchronization for new buffer allocation
+                self.allocation_tracker[(buffer_name, dtype)] = size
+                torch.cuda.synchronize()
+        return Bucket(
+            data=get_global_memory_buffer().get_tensor(
+                [size], dtype=dtype, name=buffer_name, mem_alloc_context=mem_alloc_context
+            )
+        )
+
+    def _get_gbuf_name(self, buf_group_id: int, bucket_index: int):
+        return f"{self.name}_{buf_group_id}_{bucket_index}"
+
+    def free(self, bucket_id: int):
+        """
+        free a temporary bucket.
+        """
+        fsdp_unit_id = self.fsdp_param_groups[bucket_id].fsdp_unit_id
+        if fsdp_unit_id in self.fsdp_double_buffer_units:
+            if bucket_id not in self.using_buffer:
+                # This bucket is not allocated by fixed pool allocator.
+                return
+            # Return the buffer to the idle pool.
+            self.idle_buffer.append(self.using_buffer[bucket_id])
+            del self.using_buffer[bucket_id]
+            return
+        if self.fallback_to_persistent_buffer is False:
+            # If not managed by fixed pool allocator, delegate to the backup allocator.
+            logging.debug(f"[FSDP] Free from the backup allocator for {bucket_id} {fsdp_unit_id}")
+            self.backup_allocator.free(bucket_id)
 
 
 def _get_parameter_groups(
